+++ ChangeLog.md
@@ -11,6 +11,7 @@
 - Update the example `simpleNet` to run network forward.
 - Update the example `dcgan_struct` to support GPU for training.
 - Update the example `dcgan_struct` to show more detail for usage.
+- Update the example `dcgan_struct` to convert the samples that are generated from GAN to images.
 
 ### [Fixed]
 - Fixed the dimension error for the example `dcgan_struct`.